requirements/base.txt
Django==1.6
Fabric==1.8.1
South==0.8.4
psycopg2==2.5.1
sqlparse==0.1.10
Unipath==1.0
redis==2.8.0
django-redis-sessions==0.4.0
uWSGI==2.0
django-crispy-forms==1.4.0
django-countries==2.0b4
django-bootstrap3==2.5.5
django-braces==1.3.1
django-registration==1.0
